+++ src/main/java/cart/application/CartItemService.java
@@ -1,15 +1,15 @@
 package cart.application;
 
-import cart.dao.*;
+import cart.application.dto.request.CartItemQuantityUpdateRequest;
+import cart.application.dto.request.CartItemRequest;
+import cart.application.dto.request.PaymentRequest;
+import cart.application.dto.response.CartItemResponse;
 import cart.domain.cart.CartItem;
 import cart.domain.member.Member;
 import cart.domain.order.Order;
 import cart.domain.order.OrderProducts;
 import cart.domain.product.Product;
-import cart.dto.request.CartItemQuantityUpdateRequest;
-import cart.dto.request.CartItemRequest;
-import cart.dto.request.PaymentRequest;
-import cart.dto.response.CartItemResponse;
+import cart.persistence.dao.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
