+++ src/main/java/cart/application/MemberService.java
@@ -1,9 +1,9 @@
 package cart.application;
 
-import cart.dao.MemberDao;
+import cart.application.dto.response.MemberResponse;
+import cart.application.dto.response.PointResponse;
 import cart.domain.member.Member;
-import cart.dto.response.MemberResponse;
-import cart.dto.response.PointResponse;
+import cart.persistence.dao.MemberDao;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
