+++ src/main/java/cart/application/dto/request/ProductRequest.java
@@ -1,4 +1,4 @@
-package cart.dto.request;
+package cart.application.dto.request;
 
 public class ProductRequest {
     private String name;
