+++ src/main/java/cart/persistence/dao/OrderProductDao.java
@@ -1,4 +1,4 @@
-package cart.dao;
+package cart.persistence.dao;
 
 import cart.domain.order.OrderProducts;
 import org.springframework.jdbc.core.JdbcTemplate;
