+++ src/main/java/cart/persistence/dao/ProductDao.java
@@ -1,4 +1,4 @@
-package cart.dao;
+package cart.persistence.dao;
 
 import cart.domain.product.Product;
 import org.springframework.jdbc.core.JdbcTemplate;
