+++ src/main/java/cart/ui/ProductApiController.java
@@ -1,8 +1,8 @@
 package cart.ui;
 
 import cart.application.ProductService;
-import cart.dto.request.ProductRequest;
-import cart.dto.response.ProductResponse;
+import cart.application.dto.request.ProductRequest;
+import cart.application.dto.response.ProductResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
