+++ src/test/java/cart/domain/cart/CartItemTest.java
@@ -1,6 +1,5 @@
-package cart.domain;
+package cart.domain.cart;
 
-import cart.domain.cart.CartItem;
 import cart.domain.member.Member;
 import cart.domain.product.Product;
 import cart.exception.CartItemException;
