+++ src/test/java/cart/domain/member/MemberTest.java
@@ -1,6 +1,5 @@
-package cart.domain;
+package cart.domain.member;
 
-import cart.domain.member.Member;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
