+++ src/test/java/cart/integration/ProductIntegrationTest.java
@@ -1,7 +1,7 @@
 package cart.integration;
 
-import cart.dto.request.ProductRequest;
-import cart.dto.response.ProductResponse;
+import cart.application.dto.request.ProductRequest;
+import cart.application.dto.response.ProductResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
